requirements.txt
Django==1.6
Markdown==2.3.1
South==0.8.3
argparse==1.2.1
# distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-cors-headers==0.12
django-filter==0.7
djangorestframework==2.3.9
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
flake8==2.1.0
django-oauth2-provider==0.2.6
